Moving light to its own file
